% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{rcolorUtrecht}
\alias{rcolorUtrecht}
\title{Utrecht City Colour Palletes}
\format{An object of class \code{list} of length 6.}
\usage{
rcolorUtrecht
}
\description{
Colour palletes from street art in Utrecht city. Inspired by the \link{EdwinTh}
(https://github.com/EdwinTh/dutchmasters) package.
}
\details{
Available palletes:

Het Uithof - Hogeschool Utrecht

De Strakke Hand - Fasecontrastmicroscoop, Dr. Bleeker

patatjes4life - electricity box

Jan Is De Man - Groceries above Albert Heijn

Jan Is De Man - Vogelenbuurt

De Strakke Hand - Gansstraat

Leon Keer - 3D kunst

De Strakke Hand - Ibis Hotel

Studio KCA - The Wale

Derk Wessels - Derk

Dick Bruna - Nijntje

Munir de Vries - Dichterswijk

De Verfdokter - Train Station

Boukje Lootsma - The Wonders

Anne P. Boer - Ships With Ducdalf

The Dom - Utecht University

}
\examples{
# devtools::install_github("cherjuliette/rcolorUtrecht")
# devtools::install_github("EdwinTh/dutchmasters")

gg
}
\keyword{datasets}