% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{rcolorUtrecht_pal}
\alias{rcolorUtrecht_pal}
\title{dutchmasters palette with ramped colours}
\usage{
rcolorUtrecht_pal(palette = "hu", alpha = 1, reverse = FALSE)
}
\arguments{
\item{palette}{Choose from 'rcolorUtrecht_palettes' list}

\item{alpha}{transparency}

\item{reverse}{If TRUE, the direction of the colours is reversed.}
}
\description{
dutchmasters palette with ramped colours
}
\examples{
library(scales)
show_col(rcolorUtrecht_pal("hu")(10))

filled.contour(volcano,color.palette = rcolorUtrecht_pal("miffy"), asp=1)

}