% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_color_rcolorUtrecht}
\alias{scale_color_rcolorUtrecht}
\alias{scale_color_rcolorUtrecht}
\title{Setup color palette for ggplot2}
\usage{
scale_color_rcolorUtrecht(..., palette = "hu", discrete = TRUE,
  alpha = 1, reverse = FALSE)

scale_color_rcolorUtrecht(..., palette = "hu", discrete = TRUE,
  alpha = 1, reverse = FALSE)
}
\arguments{
\item{...}{additional arguments to pass to scale_color_gradientn}

\item{palette}{Choose from 'rcolorUtrecht_palettes' list}

\item{discrete}{whether to use a discrete color palette}

\item{alpha}{transparency}

\item{reverse}{logical, Reverse the order of the colors?}
}
\description{
Setup color palette for ggplot2
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(mpg, wt)) +
  geom_point(aes(color = factor(cyl)), size = 4) +
  scale_color_rcolorUtrecht()
ggplot(mtcars, aes(mpg, wt)) +
  geom_point(aes(color = hp)) +
  scale_color_rcolorUtrecht(palette="miffy", discrete = FALSE)
ggplot(data = mpg) +
  geom_point(mapping = aes(x = displ, y = hwy, color = class)) +
  scale_color_rcolorUtrecht(palette="uu")
ggplot(diamonds) + geom_bar(aes(x = cut, fill = clarity)) +
  scale_fill_rcolorUtrecht(palette = "the_wonders")
}